+++ CHANGELOG.md
@@ -1,10 +1,12 @@
 # Revision history for ghci-quickfix
 
+## 0.1.1.1 -- 2026-07-14
+* Sort errors in quickfix file by order of emission
+
 ## 0.1.1.0 -- 2026-05-13
 
 * Support GHC 9.14
 * Drop support for 9.6
-* Sort errors in quickfix file by order of emission
 
 ## 0.1.0.0 -- 2026-01-09